A simple UIAlertActionSheet should do the job. I think?
Implement a swipe recogniser (.Up) and check for the location in TouchesMoved Method. We need the location movement about the Y axis so, let's say it the user moves the view 20 points upwards and the swipe direction being up while peeking, bring in an alert view in the action sheet style displaying the options needed. My implementation might be wrong but do let me know.